// typescript-utils/nightly-nightly-digital-aura-analyze/src/types.ts
export enum DigitalAura {
  AmbitiousAscent = "Ambitious Ascent",
  SereneScroll = "Serene Scroll",
  FleetingWhisper = "Fleeting Whisper",
  MysteriousMuddle = "Mysterious Muddle",
  ReflectiveArchive = "Reflective Archive",
  ChaoticCascade = "Chaotic Cascade",
  HarmoniousHub = "Harmonious Hub",
  SilentSentinel = "Silent Sentinel",
  VibrantVenture = "Vibrant Venture",
  EphemeralEcho = "Ephemeral Echo",
  ForgottenFragment = "Forgotten Fragment"
}

export type KeywordMap = {
  [key: string]: DigitalAura;
};